Improve how we record and handle entity_history
The main focus of this change is to stop recording entity_history for the
'subgraphs' subgraph, and to avoid recording history for changes that
actually do not change the entity. The history we recorded for these made
up about 75% of the storage needed in the hosted offering, and we never
need the data thus recorded.

This change also simplifies how the revert_block stored procedure works by
basing it off a single history query, and removes unused input parameters
for it.

The change also gets rid of a few stored procedures that were previously
used, but are not any longer.

Note that this change also subtly alters the StoreEvents sent during a
revert: before the entity change(s) and the change to the subgraphs
metadata were delivered in one event, where they are now sent as two
separate events.

More background on these changes can be found at
https://github.com/graphprotocol/graph-node/issues/807